requirement.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
backports.functools-lru-cache==2.0.0
black==25.1.0
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comtypes==1.4.11
dataclasses-json==0.6.7
deepgram-sdk==2.11.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
elevenlabs==2.8.1
fastapi==0.116.1
filelock==3.18.0
flake8==7.3.0
frozenlist==1.7.0
fsspec==2025.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.30.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
indic-transliteration==2.3.69
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
keyboard==0.13.5
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.72
langchain-groq==0.3.6
langchain-huggingface==0.3.1
langchain-mongodb==0.6.2
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.5
lark==1.2.2
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai==1.95.1
openai-whisper==20250625
orjson==3.10.18
packaging==24.2
pathspec==0.12.1
pip==25.1.1
platformdirs==4.3.8
playsound==1.2.2
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
pydub==0.25.1
pyflakes==3.4.0
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.13.2
pypiwin32==223
pytest==8.4.1
python-dotenv==1.1.1
pyttsx3==2.99
pywin32==311
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
roman==5.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
SQLAlchemy==2.0.41
starlette==0.47.1
structlog==25.4.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.4
toml==0.10.2
torch==2.7.1
tqdm==4.67.1
twilio==9.6.5
typer==0.16.0
typing_extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
    websockets==12.0
yarl==1.20.1
zstandard==0.23.0
